Rework test utilities and more work on activity workers
